requirements.txt
google-genai
python-dotenv
flask
flask-session
flask-cors
pymongo